pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "nw-watch"
version = "0.1.0"
description = "Dual-device network CLI monitor"
requires-python = ">=3.11"
dependencies = [
    "netmiko>=4.3.0",
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "jinja2>=3.1.3",
    "pyyaml>=6.0.1",
    "aiofiles>=23.2.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.26.0",
]

[tool.setuptools.packages.find]
include = ["collector*", "webapp*", "shared*"]
exclude = ["tests*", "data*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
